Put VaultSecret in vault.py; delete dataclasses.py
